Add Dockerfile and README for ZeroTier containerization
This commit includes a Dockerfile and README.md for creating a lightweight
ZeroTier container based on Alpine Linux. The Dockerfile outlines the
build process, including the installation of necessary packages and
compilation of the ZeroTierOne project. The README provides instructions
for running the container and joining ZeroTier networks.

The container is designed to run ZeroTier One as a service on
container-oriented distributions like Fedora CoreOS, but can be used
on any Linux system with Docker or Podman.