Add Dockerfile and README for Cloudflare Warp container
This commit introduces the Dockerfile and README.md necessary to create
a Cloudflare Warp container. The Dockerfile specifies the build process,
including the installation of required packages and Cloudflare WARP.
The README provides instructions for running the container and configuring
WARP settings such as the license key and families mode.

An entrypoint.sh script is also included to manage the WARP connection
and socat is used to redirect traffic from port 1080 to 40000.