Add Dockerfile and README for vlmcsd KMS server container
This commit adds a Dockerfile and README.md for setting up a fully Microsoft
compatible Key Management Service (KMS) server using vlmcsd. The Dockerfile
specifies the build process, including the installation of required packages
and the compilation of vlmcsd from source. The README provides instructions
for running the container and exposes the KMS port.

The vlmcsd server is designed to be a drop-in replacement for a Microsoft KMS
server and supports KMS protocol versions 4, 5, and 6. It is suitable for
embedded systems and POSIX compatible operating systems.