feat(ci): add support for Aliyun Container Registry login and update Docker push logic
Add login to Aliyun Container Registry for authentication with Alibaba Cloud. This step
is crucial for deployment and pushing Docker images to the Alibaba Cloud Container
Registry. Also, update the Docker push logic to align with the new multi-platform support
and tagging requirements, including pushes to both Docker Hub and Aliyun Container
Registry.